ProbABEL: change back to using doubles instead of floats. With floats reading from text files worked, but reading from filevector files gave rounding errors (even though the fv files were of type FLOAT). Now I've replaced the conversion that was present in v0.4.2 from C++ style to C style and this seems to work.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/ProbABEL@1540 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/coxph_data.cpp
@@ -125,7 +125,7 @@ coxph_data::coxph_data(phedata &phed, gendata &gend, const int snpnum)
     {
         for (int j = 0; j < ngpreds; j++)
         {
-            float *snpdata = new float[nids];
+            double *snpdata = new double[nids];
             gend.get_var(snpnum * ngpreds + j, snpdata);
             for (int i = 0; i < nids; i++)
             {
@@ -216,7 +216,7 @@ void coxph_data::update_snp(gendata *gend, const int snpnum) {
     freq   = 0.0;
 
     for (int j = 0; j < ngpreds; j++) {
-        float *snpdata = new float[nids];
+        double *snpdata = new double[nids];
         for (int i = 0; i < nids; i++) {
             masked_data[i] = 0;
         }
